chore: move tag input and combo select to input components, remove duplicated icon
