--- src/components/ComboSelect.css
@@ -1,29 +1,0 @@
-.combo-select {
-  width: 100%;
-  display: block;
-  box-sizing: border-box;
-  position: relative;
-  overflow: hidden;
-}
-
-.combo-select input {
-  padding: 26px 2px 12px 16px;
-  width: 100%;
-  box-sizing: border-box;
-  border-radius: 4px;
-  border: 1px solid rgb(176, 176, 176);
-  transition: all 0.2s ease-in;
-  outline: none;
-}
-
-.combo-select .button-sheet {
-  background: white;
-  position: absolute;
-  bottom: 2px;
-  right: 2px;
-  width: 40px;
-  height: 40px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
