+++ src/components/Input/Input.css
@@ -69,3 +69,33 @@
   padding: 8px;
   background: rgb(239, 239, 239);
 }
+
+.combo-select {
+  width: 100%;
+  display: block;
+  box-sizing: border-box;
+  position: relative;
+  overflow: hidden;
+}
+
+.combo-select input {
+  padding: 26px 2px 12px 16px;
+  width: 100%;
+  box-sizing: border-box;
+  border-radius: 4px;
+  border: 1px solid rgb(176, 176, 176);
+  transition: all 0.2s ease-in;
+  outline: none;
+}
+
+.combo-select .button-sheet {
+  background: white;
+  position: absolute;
+  bottom: 2px;
+  right: 2px;
+  width: 40px;
+  height: 40px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
