+++ src/components/Input/index.ts
@@ -1,0 +1,3 @@
+export { Input } from './Input';
+export { TagInput } from './TagInput';
+export { ComboSelect } from './ComboSelect';
