+++ src/components/Message/index.tsx
@@ -7,7 +7,7 @@ import { CSSTransition } from 'react-transition-group';
 import { Button } from '../Button';
 import ErrorIcon from '../icons/Error';
 import SuccessIcon from '../icons/Success';
-import XIcon from '../icons/X';
+import CloseIcon from '../icons/Close';
 
 export default function Message({ data, onClose }: { onClose: () => void; data: AppMessage }) {
   const [open, setOpen] = useState<boolean>(false);
@@ -45,7 +45,7 @@ export default function Message({ data, onClose }: { onClose: () => void; data: 
         {iconEl}
         <p>{data.message}</p>
         <Button onClick={onClose}>
-          <XIcon />
+          <CloseIcon />
         </Button>
       </article>
     </CSSTransition>
